+++ tests/scripts/shared/test_script_syntax.py
@@ -3,7 +3,7 @@
 
 Tests scripts for syntax errors
     - Tests .py, .R and .sh scripts
-    - Test fails if any scripts have other file extenstions
+    - Test fails if any scripts have other file extensions
 
 Authors:
     - Cameron Fraser
@@ -12,9 +12,8 @@ Authors:
 # Import libraries
 import subprocess
 import pytest
-from pathlib import Path
 import pytest
-from conftest import PROJECT_ROOT
+from tests.conftest import PROJECT_ROOT
 
 # Pytest marking
 pytestmark = [